+++ packages/rabbitmq/src/amqp/connection.ts
@@ -30,7 +30,6 @@ import {
   RabbitMQChannelConfig,
   ConsumeOptions,
   MessageDeserializer,
-  MessageSerializer,
 } from '../rabbitmq.interfaces';
 import {
   getHandlerForLegacyBehavior,
@@ -47,8 +46,20 @@ export type ConsumerTag = string;
 export type SubscriberHandler<T = unknown> = (
   msg: T | undefined,
   rawMessage?: ConsumeMessage,
-  headers?: any
+  headers?: any,
 ) => Promise<SubscribeResponse>;
+
+export type BatchSubscriberHandler<T = unknown> = (
+  msg: (T | undefined)[],
+  rawMessage?: ConsumeMessage[],
+  headers?: any[],
+) => Promise<SubscribeResponse>;
+
+export type RpcSubscriberHandler<T = unknown, U = unknown> = (
+  msg: T | undefined,
+  rawMessage?: ConsumeMessage,
+  headers?: any,
+) => Promise<RpcResponse<U>>;
 
 export interface CorrelationMessage {
   correlationId: string;
@@ -63,26 +74,21 @@ export interface SubscriptionResult {
 export type BaseConsumerHandler = {
   consumerTag: string;
   channel: ConfirmChannel;
+  msgOptions: MessageHandlerOptions;
 };
 
 export type ConsumerHandler<T, U> =
   | (BaseConsumerHandler & {
       type: 'subscribe';
-      msgOptions: MessageHandlerOptions;
-      handler: (
-        msg: T | undefined,
-        rawMessage?: ConsumeMessage,
-        headers?: any
-      ) => Promise<SubscribeResponse>;
+      handler: SubscriberHandler<T>;
+    })
+  | (BaseConsumerHandler & {
+      type: 'subscribe-batch';
+      handler: BatchSubscriberHandler<T>;
     })
   | (BaseConsumerHandler & {
       type: 'rpc';
-      rpcOptions: MessageHandlerOptions;
-      handler: (
-        msg: T | undefined,
-        rawMessage?: ConsumeMessage,
-        headers?: any
-      ) => Promise<RpcResponse<U>>;
+      handler: RpcSubscriberHandler<T, U>;
     });
 
 type Consumer = (msg: ConsumeMessage | null) => void | Promise<void>;
@@ -92,7 +98,7 @@ const defaultConfig = {
   prefetchCount: 10,
   defaultExchangeType: 'topic',
   defaultRpcErrorHandler: getHandlerForLegacyBehavior(
-    MessageHandlerErrorBehavior.REQUEUE
+    MessageHandlerErrorBehavior.REQUEUE,
   ),
   defaultSubscribeErrorBehavior: MessageHandlerErrorBehavior.REQUEUE,
   exchanges: [],
@@ -109,6 +115,7 @@ const defaultConfig = {
   enableDirectReplyTo: true,
   channels: {},
   handlers: {},
+  defaultHandler: '',
   enableControllerDiscovery: false,
 };
 
@@ -140,7 +147,7 @@ export class AmqpConnection {
     this.config = {
       deserializer: (message) => JSON.parse(message.toString()),
       serializer: (value) => Buffer.from(JSON.stringify(value)),
-      logger: config.logger || new Logger(AmqpConnection.name),
+      logger: config?.logger || new Logger(AmqpConnection.name),
       ...defaultConfig,
       ...config,
     };
@@ -202,43 +209,43 @@ export class AmqpConnection {
             throwError(
               () =>
                 new Error(
-                  `Failed to connect to a RabbitMQ broker within a timeout of ${timeoutInterval}ms`
-                )
+                  `Failed to connect to a RabbitMQ broker within a timeout of ${timeoutInterval}ms`,
+                ),
             ),
         }),
-        catchError((err) => (reject ? throwError(() => err) : EMPTY))
-      )
+        catchError((err) => (reject ? throwError(() => err) : EMPTY)),
+      ),
     );
   }
 
   private async initCore(): Promise<void> {
     this.logger.log(
-      `Trying to connect to RabbitMQ broker (${this.config.name})`
+      `Trying to connect to RabbitMQ broker (${this.config.name})`,
     );
 
     this._managedConnection = connect(
       Array.isArray(this.config.uri) ? this.config.uri : [this.config.uri],
-      this.config.connectionManagerOptions
+      this.config.connectionManagerOptions,
     );
 
     this._managedConnection.on('connect', ({ connection }) => {
       this._connection = connection;
       this.logger.log(
-        `Successfully connected to RabbitMQ broker (${this.config.name})`
+        `Successfully connected to RabbitMQ broker (${this.config.name})`,
       );
     });
 
     this._managedConnection.on('disconnect', ({ err }) => {
       this.logger.error(
         `Disconnected from RabbitMQ broker (${this.config.name})`,
-        err?.stack
+        err?.stack,
       );
     });
 
     this._managedConnection.on('connectFailed', ({ err }) => {
       this.logger.error(
         `Failed to connect to RabbitMQ broker (${this.config.name})`,
-        err?.stack
+        err?.stack,
       );
     });
 
@@ -266,7 +273,7 @@ export class AmqpConnection {
             ...config,
             default: false,
           });
-        })
+        }),
       ),
       this.setupManagedChannel(defaultChannel.name, defaultChannel.config),
     ]);
@@ -275,7 +282,7 @@ export class AmqpConnection {
   private async setupInitChannel(
     channel: ConfirmChannel,
     name: string,
-    config: RabbitMQChannelConfig
+    config: RabbitMQChannelConfig,
   ): Promise<void> {
     this._channels[name] = channel;
 
@@ -293,11 +300,11 @@ export class AmqpConnection {
             return channel.assertExchange(
               x.name,
               x.type || this.config.defaultExchangeType,
-              x.options
+              x.options,
             );
           }
           return channel.checkExchange(x.name);
-        })
+        }),
       );
 
       await Promise.all(
@@ -306,9 +313,9 @@ export class AmqpConnection {
             exchangeBinding.destination,
             exchangeBinding.source,
             exchangeBinding.pattern,
-            exchangeBinding.args
-          )
-        )
+            exchangeBinding.args,
+          ),
+        ),
       );
 
       await this.setupQueuesWithBindings(channel, this.config.queues);
@@ -323,7 +330,7 @@ export class AmqpConnection {
 
   private async setupQueuesWithBindings(
     channel: ConfirmChannel,
-    queues: RabbitMQQueueConfig[]
+    queues: RabbitMQQueueConfig[],
   ) {
     await Promise.all(
       queues.map(async (configuredQueue) => {
@@ -339,9 +346,9 @@ export class AmqpConnection {
             ...(name !== undefined && { queue: name }),
             queueOptions,
           },
-          channel
+          channel,
         );
-      })
+      }),
     );
   }
 
@@ -370,7 +377,7 @@ export class AmqpConnection {
       },
       {
         noAck: true,
-      }
+      },
     );
   }
 
@@ -384,19 +391,19 @@ export class AmqpConnection {
       filter((x) =>
         requestId
           ? x.correlationId === correlationId && x.requestId === requestId
-          : x.correlationId === correlationId
+          : x.correlationId === correlationId,
       ),
       map((x) => x.message as T),
-      first()
+      first(),
     );
 
     const timeout$ = interval(timeout).pipe(
       first(),
       map(() => {
         throw new Error(
-          `Failed to receive response within timeout of ${timeout}ms for exchange "${requestOptions.exchange}" and routing key "${requestOptions.routingKey}"`
+          `Failed to receive response within timeout of ${timeout}ms for exchange "${requestOptions.exchange}" and routing key "${requestOptions.routingKey}"`,
         );
-      })
+      }),
     );
 
     const result = lastValueFrom(race(response$, timeout$));
@@ -411,7 +418,7 @@ export class AmqpConnection {
         correlationId,
         headers: requestOptions.headers,
         expiration: requestOptions.expiration,
-      }
+      },
     );
 
     return result;
@@ -421,20 +428,44 @@ export class AmqpConnection {
     handler: SubscriberHandler<T>,
     msgOptions: MessageHandlerOptions,
     originalHandlerName: string,
-    consumeOptions?: ConsumeOptions
+    consumeOptions?: ConsumeOptions,
+  ): Promise<SubscriptionResult> {
+    return this.consumerFactory(msgOptions, (channel) =>
+      this.setupSubscriberChannel<T>(
+        handler,
+        msgOptions,
+        channel,
+        originalHandlerName,
+        consumeOptions,
+      ),
+    );
+  }
+
+  public async createBatchSubscriber<T>(
+    handler: BatchSubscriberHandler<T>,
+    msgOptions: MessageHandlerOptions,
+    consumeOptions?: ConsumeOptions,
+  ): Promise<SubscriptionResult> {
+    return this.consumerFactory(msgOptions, (channel) =>
+      this.setupBatchSubscriberChannel<T>(
+        handler,
+        msgOptions,
+        channel,
+        consumeOptions,
+      ),
+    );
+  }
+
+  private async consumerFactory(
+    msgOptions: MessageHandlerOptions,
+    setupFunction: (channel: ConfirmChannel) => Promise<string>,
   ): Promise<SubscriptionResult> {
     return new Promise((res) => {
       this.selectManagedChannel(msgOptions?.queueOptions?.channel).addSetup(
-        async (channel) => {
-          const consumerTag = await this.setupSubscriberChannel<T>(
-            handler,
-            msgOptions,
-            channel,
-            originalHandlerName,
-            consumeOptions
-          );
+        async (channel: ConfirmChannel) => {
+          const consumerTag = await setupFunction(channel);
           res({ consumerTag });
-        }
+        },
       );
     });
   }
@@ -448,7 +479,7 @@ export class AmqpConnection {
       const messageProcessingPromise = Promise.resolve(consumer(msg));
       this.outstandingMessageProcessing.add(messageProcessingPromise);
       messageProcessingPromise.finally(() =>
-        this.outstandingMessageProcessing.delete(messageProcessingPromise)
+        this.outstandingMessageProcessing.delete(messageProcessingPromise),
       );
     };
   }
@@ -458,7 +489,7 @@ export class AmqpConnection {
     msgOptions: MessageHandlerOptions,
     channel: ConfirmChannel,
     originalHandlerName = 'unknown',
-    consumeOptions?: ConsumeOptions
+    consumeOptions?: ConsumeOptions,
   ): Promise<ConsumerTag> {
     const queue = await this.setupQueue(msgOptions, channel);
 
@@ -470,10 +501,8 @@ export class AmqpConnection {
             throw new Error('Received null message');
           }
 
-          const response = await this.handleMessage(handler, msg, {
-            allowNonJsonMessages: msgOptions.allowNonJsonMessages,
-            deserializer: msgOptions.deserializer,
-          });
+          const result = this.deserializeMessage<T>(msg, msgOptions);
+          const response = await handler(result.message, msg, result.headers);
 
           if (response instanceof Nack) {
             channel.nack(msg, false, response.requeue);
@@ -485,8 +514,8 @@ export class AmqpConnection {
           if (response) {
             this.logger.warn(
               `Received response: [${this.config.serializer(
-                response
-              )}] from subscribe handler [${originalHandlerName}]. Subscribe handlers should only return void`
+                response,
+              )}] from subscribe handler [${originalHandlerName}]. Subscribe handlers should only return void`,
             );
           }
 
@@ -499,14 +528,14 @@ export class AmqpConnection {
               msgOptions.errorHandler ||
               getHandlerForLegacyBehavior(
                 msgOptions.errorBehavior ||
-                  this.config.defaultSubscribeErrorBehavior
+                  this.config.defaultSubscribeErrorBehavior,
               );
 
             await errorHandler(channel, msg, e);
           }
         }
       }),
-      consumeOptions
+      consumeOptions,
     );
 
     this.registerConsumerForQueue({
@@ -520,36 +549,145 @@ export class AmqpConnection {
     return consumerTag;
   }
 
-  public async createRpc<T, U>(
-    handler: (
-      msg: T | undefined,
-      rawMessage?: ConsumeMessage,
-      headers?: any
-    ) => Promise<RpcResponse<U>>,
-    rpcOptions: MessageHandlerOptions
-  ): Promise<SubscriptionResult> {
-    return new Promise((res) => {
-      this.selectManagedChannel(rpcOptions?.queueOptions?.channel).addSetup(
-        async (channel) => {
-          const consumerTag = await this.setupRpcChannel<T, U>(
-            handler,
-            rpcOptions,
-            channel
-          );
-          res({ consumerTag });
-        }
+  private async setupBatchSubscriberChannel<T>(
+    handler: BatchSubscriberHandler<T>,
+    msgOptions: MessageHandlerOptions,
+    channel: ConfirmChannel,
+    consumeOptions?: ConsumeOptions,
+  ): Promise<ConsumerTag> {
+    let batchSize = msgOptions.batchOptions?.size;
+    let batchTimeout = msgOptions.batchOptions?.timeout;
+    let batchMsgs: ConsumeMessage[] = [];
+    let batchTimer: NodeJS.Timeout;
+    let inflightBatchHandler: () => Promise<void>;
+
+    // Normalize batch values but warn consumer about this adjusts
+    if (!batchSize || batchSize < 2) {
+      this.logger.warn(
+        `batch size too low/not defined, received: ${batchSize}. Adjusting to 10`,
       );
+      batchSize = 10;
+    }
+
+    if (!batchTimeout || batchTimeout < 1) {
+      this.logger.warn(
+        `batch timeout too low/not defined, received: ${batchTimeout}. Setting timeout to 200ms`,
+      );
+      batchTimeout = 200;
+    }
+
+    const queue = await this.setupQueue(msgOptions, channel);
+
+    const { consumerTag }: { consumerTag: ConsumerTag } = await channel.consume(
+      queue,
+      this.wrapConsumer(async (msg) => {
+        if (isNull(msg)) {
+          return;
+        }
+
+        batchMsgs.push(msg);
+
+        if (batchMsgs.length === 1) {
+          // Wrapped in a Promise to ensure outstanding message logic is aware.
+          await new Promise<void>((resolve) => {
+            const batchHandler = async () => {
+              const batchMsgsToProcess = batchMsgs;
+              batchMsgs = [];
+
+              await this.handleBatchedMessages(
+                handler,
+                msgOptions,
+                channel,
+                batchMsgsToProcess,
+              );
+
+              resolve();
+            };
+
+            batchTimer = setTimeout(batchHandler, batchTimeout);
+            inflightBatchHandler = batchHandler;
+          });
+        } else if (batchMsgs.length === batchSize) {
+          clearTimeout(batchTimer);
+          await inflightBatchHandler();
+        } else {
+          batchTimer.refresh();
+        }
+      }),
+      consumeOptions,
+    );
+
+    this.registerConsumerForQueue({
+      type: 'subscribe-batch',
+      consumerTag,
+      handler,
+      msgOptions,
+      channel,
     });
+
+    return consumerTag;
+  }
+
+  private async handleBatchedMessages<T>(
+    handler: BatchSubscriberHandler<T>,
+    msgOptions: MessageHandlerOptions,
+    channel: ConfirmChannel,
+    batchMsgs: ConsumeMessage[],
+  ) {
+    try {
+      const messages: (T | undefined)[] = [];
+      const headers: any[] = [];
+
+      for (const msg of batchMsgs) {
+        const result = this.deserializeMessage<T>(msg, msgOptions);
+        messages.push(result.message);
+        headers.push(result.headers);
+      }
+
+      const response = await handler(messages, batchMsgs, headers);
+
+      if (response instanceof Nack) {
+        for (const msg of batchMsgs) {
+          channel.nack(msg, false, response.requeue);
+        }
+        return;
+      }
+
+      for (const msg of batchMsgs) {
+        channel.ack(msg);
+      }
+    } catch (e) {
+      const batchErrorHandler = msgOptions.batchOptions?.errorHandler;
+      const errorHandler = msgOptions.errorHandler;
+      const defaultErrorHandler = getHandlerForLegacyBehavior(
+        msgOptions.errorBehavior || this.config.defaultSubscribeErrorBehavior,
+      );
+
+      if (batchErrorHandler) {
+        await batchErrorHandler(channel, batchMsgs, e);
+      } else if (errorHandler) {
+        for (const msg of batchMsgs) {
+          await errorHandler(channel, msg, e);
+        }
+      } else {
+        await defaultErrorHandler(channel, batchMsgs, e);
+      }
+    }
+  }
+
+  public async createRpc<T, U>(
+    handler: RpcSubscriberHandler<T, U>,
+    rpcOptions: MessageHandlerOptions,
+  ): Promise<SubscriptionResult> {
+    return this.consumerFactory(rpcOptions, (channel) =>
+      this.setupRpcChannel<T, U>(handler, rpcOptions, channel),
+    );
   }
 
   public async setupRpcChannel<T, U>(
-    handler: (
-      msg: T | undefined,
-      rawMessage?: ConsumeMessage,
-      headers?: any
-    ) => Promise<RpcResponse<U>>,
+    handler: RpcSubscriberHandler<T, U>,
     rpcOptions: MessageHandlerOptions,
-    channel: ConfirmChannel
+    channel: ConfirmChannel,
   ): Promise<ConsumerTag> {
     const queue = await this.setupQueue(rpcOptions, channel);
 
@@ -567,15 +705,13 @@ export class AmqpConnection {
             channel.nack(msg, false, false);
             this.logger.error(
               'Received message with invalid routing key: ' +
-                msg.fields.routingKey
+                msg.fields.routingKey,
             );
             return;
           }
 
-          const response = await this.handleMessage(handler, msg, {
-            allowNonJsonMessages: rpcOptions.allowNonJsonMessages,
-            deserializer: rpcOptions.deserializer,
-          });
+          const result = this.deserializeMessage<T>(msg, rpcOptions);
+          const response = await handler(result.message, msg, result.headers);
 
           if (response instanceof Nack) {
             channel.nack(msg, false, response.requeue);
@@ -602,32 +738,32 @@ export class AmqpConnection {
               this.config.defaultRpcErrorHandler ||
               getHandlerForLegacyBehavior(
                 rpcOptions.errorBehavior ||
-                  this.config.defaultSubscribeErrorBehavior
+                  this.config.defaultSubscribeErrorBehavior,
               );
 
             await errorHandler(channel, msg, e);
           }
         }
       }),
-      rpcOptions?.queueOptions?.consumerOptions
+      rpcOptions?.queueOptions?.consumerOptions,
     );
 
     this.registerConsumerForQueue({
       type: 'rpc',
       consumerTag,
       handler,
-      rpcOptions,
+      msgOptions: rpcOptions,
       channel,
     });
 
     return consumerTag;
   }
 
-  public publish<T = any>(
+  public publish(
     exchange: string,
     routingKey: string,
-    message: T,
-    options?: Options.Publish
+    message: any,
+    options?: Options.Publish,
   ): Promise<boolean> {
     let buffer: Buffer;
     if (message instanceof Buffer) {
@@ -643,17 +779,12 @@ export class AmqpConnection {
     return this._managedChannel.publish(exchange, routingKey, buffer, options);
   }
 
-  private handleMessage<T, U>(
-    handler: (
-      msg: T | undefined,
-      rawMessage?: ConsumeMessage,
-      headers?: any
-    ) => Promise<U>,
+  private deserializeMessage<T>(
     msg: ConsumeMessage,
     options: {
       allowNonJsonMessages?: boolean;
       deserializer?: MessageDeserializer;
-    }
+    },
   ) {
     let message: T | undefined = undefined;
     let headers: any = undefined;
@@ -676,12 +807,12 @@ export class AmqpConnection {
       headers = msg.properties.headers;
     }
 
-    return handler(message, msg, headers);
+    return { message, headers };
   }
 
   private async setupQueue(
     subscriptionOptions: MessageHandlerOptions,
-    channel: ConfirmChannel
+    channel: ConfirmChannel,
   ): Promise<string> {
     const {
       exchange,
@@ -703,12 +834,12 @@ export class AmqpConnection {
           channel,
           queueName,
           queueOptions,
-          error
+          error,
         );
       }
     } else {
       const { queue } = await channel.checkQueue(
-        subscriptionOptions.queue || ''
+        subscriptionOptions.queue || '',
       );
       actualQueue = queue;
     }
@@ -728,10 +859,10 @@ export class AmqpConnection {
               actualQueue as string,
               exchange,
               routingKey,
-              bindQueueArguments
+              bindQueueArguments,
             );
           }
-        })
+        }),
       );
     }
 
@@ -750,17 +881,17 @@ export class AmqpConnection {
     }
 
     channel.on('connect', () =>
-      this.logger.log(`Successfully connected a RabbitMQ channel "${name}"`)
+      this.logger.log(`Successfully connected a RabbitMQ channel "${name}"`),
     );
 
     channel.on('error', (err, { name }) =>
-      this.logger.log(
-        `Failed to setup a RabbitMQ channel - name: ${name} / error: ${err.message} ${err.stack}`
-      )
+      this.logger.error(
+        `Failed to setup a RabbitMQ channel - name: ${name} / error: ${err.message} ${err.stack}`,
+      ),
     );
 
     channel.on('close', () =>
-      this.logger.log(`Successfully closed a RabbitMQ channel "${name}"`)
+      this.logger.log(`Successfully closed a RabbitMQ channel "${name}"`),
     );
 
     return channel.addSetup((c) => this.setupInitChannel(c, name, config));
@@ -776,7 +907,7 @@ export class AmqpConnection {
     const channel = this._managedChannels[name];
     if (!channel) {
       this.logger.warn(
-        `Channel "${name}" does not exist, using default channel.`
+        `Channel "${name}" does not exist, using default channel: ${this._managedChannel.name}.`,
       );
 
       return this._managedChannel;
@@ -805,12 +936,13 @@ export class AmqpConnection {
   public async cancelConsumer(consumerTag: ConsumerTag) {
     const consumer = this.getConsumer(consumerTag);
     if (consumer && consumer.channel) {
+      this.logger.log(`Canceling consumer with tag: ${consumerTag}`);
       await consumer.channel.cancel(consumerTag);
     }
   }
 
   public async resumeConsumer<T, U>(
-    consumerTag: ConsumerTag
+    consumerTag: ConsumerTag,
   ): Promise<ConsumerTag | null> {
     const consumer = this.getConsumer(consumerTag) as ConsumerHandler<T, U>;
     if (!consumer) {
@@ -820,14 +952,26 @@ export class AmqpConnection {
     if (consumer.type === 'rpc') {
       newConsumerTag = await this.setupRpcChannel<T, U>(
         consumer.handler,
-        consumer.rpcOptions,
-        consumer.channel
+        consumer.msgOptions,
+        consumer.channel,
       );
-    } else {
+    } else if (consumer.type === 'subscribe') {
       newConsumerTag = await this.setupSubscriberChannel<T>(
         consumer.handler,
         consumer.msgOptions,
-        consumer.channel
+        consumer.channel,
+      );
+    } else if (consumer.type === 'subscribe-batch') {
+      newConsumerTag = await this.setupBatchSubscriberChannel<T>(
+        consumer.handler,
+        consumer.msgOptions,
+        consumer.channel,
+      );
+    } else {
+      throw new Error(
+        `Unable to resume consumer tag ${consumerTag}. Unexpected consumer type ${
+          (consumer as any).type
+        }.`,
       );
     }
     // A new consumerTag was created, remove old
@@ -844,8 +988,7 @@ export class AmqpConnection {
     // Wait for all the outstanding messages to be processed
     if (this.outstandingMessageProcessing.size) {
       this.logger.log(
-        { outstandingMessageCount: this.outstandingMessageProcessing.size },
-        'Waiting for outstanding consumers'
+        `Waiting for outstanding consumers, outstanding message count: ${this.outstandingMessageProcessing.size}`,
       );
     }
     await Promise.all(this.outstandingMessageProcessing);