+++ packages/rabbitmq/src/rabbitmq.interfaces.ts
@@ -4,6 +4,7 @@ import { ConsumeMessage, Options } from 'amqplib';
 import {
   AssertQueueErrorHandler,
   MessageErrorHandler,
+  BatchMessageErrorHandler,
   MessageHandlerErrorBehavior,
 } from './amqp/errorBehaviors';
 
@@ -80,7 +81,7 @@ export type MessageSerializer = (value: any) => Buffer;
 
 export interface MessageHandlerOptions {
   /**
-   * You can use a handler config specificied in module level.
+   * You can use a handler config specified in module level.
    * Just use the same key name defined there.
    */
   name?: string;
@@ -116,6 +117,11 @@ export interface MessageHandlerOptions {
    * If set, will override the module's default deserializer.
    */
   deserializer?: MessageDeserializer;
+
+  /**
+   * Enables consumer-side batching.
+   */
+  batchOptions?: BatchOptions;
 }
 
 export interface ConnectionInitOptions {
@@ -162,6 +168,12 @@ export interface RabbitMQConfig {
   handlers?: RabbitMQHandlers;
 
   /**
+   * You can set this property to define the default handler configuration to use
+   * when using handlers.
+   */
+  defaultHandler?: string;
+
+  /**
    * You can pass your implementation of the Nestjs LoggerService.
    */
   logger?: LoggerService;
@@ -194,4 +206,31 @@ export interface RabbitMQChannelConfig {
    * If no channel has been marked as default, new channel will be created.
    */
   default?: boolean;
+}
+
+interface BatchOptions {
+  /**
+   * The number of messages to accumulate before calling the message handler.
+   *
+   * This should be smaller than the channel prefetch.
+   *
+   * Defaults to 10 if provided value is less than 2.
+   *
+   * @default 10
+   */
+  size: number;
+
+  /**
+   * The time to wait, in milliseconds, for additional messages before returning a partial batch.
+   *
+   * Defaults to 200 if not provided or provided value is less than 1.
+   *
+   * @default 200
+   */
+  timeout?: number;
+
+  /**
+   * A function that will be called if an error is thrown during processing of an incoming message
+   */
+  errorHandler?: BatchMessageErrorHandler;
 }