+++ packages/rabbitmq/src/rabbitmq.module.ts
@@ -3,10 +3,7 @@ import {
   DiscoveryModule,
   DiscoveryService,
 } from '@golevelup/nestjs-discovery';
-import {
-  createConfigurableDynamicRootModule,
-  IConfigurableDynamicRootModule,
-} from '@golevelup/nestjs-modules';
+import { createConfigurableDynamicRootModule } from '@golevelup/nestjs-modules';
 import {
   DynamicModule,
   Module,
@@ -21,14 +18,14 @@ import { ExternalContextCreator } from '@nestjs/core/helpers/external-context-cr
 import { groupBy } from 'lodash';
 import { AmqpConnection } from './amqp/connection';
 import { AmqpConnectionManager } from './amqp/connectionManager';
-import { RABBIT_CONFIG_TOKEN, RABBIT_HANDLER } from './rabbitmq.constants';
+import { assertRabbitMqUri } from './amqp/utils';
+import {
+  RABBIT_CONFIG_TOKEN,
+  RABBIT_CONTEXT_TYPE_KEY,
+  RABBIT_HANDLER,
+} from './rabbitmq.constants';
 import { RabbitRpcParamsFactory } from './rabbitmq.factory';
 import { RabbitHandlerConfig, RabbitMQConfig } from './rabbitmq.interfaces';
-
-declare const placeholder: IConfigurableDynamicRootModule<
-  RabbitMQModule,
-  RabbitMQConfig
->;
 
 @Module({
   imports: [DiscoveryModule],
@@ -41,7 +38,7 @@ export class RabbitMQModule
         {
           provide: AmqpConnectionManager,
           useFactory: async (
-            config: RabbitMQConfig
+            config: RabbitMQConfig,
           ): Promise<AmqpConnectionManager> => {
             await RabbitMQModule.AmqpConnectionFactory(config);
             return RabbitMQModule.connectionManager;
@@ -52,10 +49,10 @@ export class RabbitMQModule
           provide: AmqpConnection,
           useFactory: async (
             config: RabbitMQConfig,
-            connectionManager: AmqpConnectionManager
+            connectionManager: AmqpConnectionManager,
           ): Promise<AmqpConnection> => {
             return connectionManager.getConnection(
-              config?.name || 'default'
+              config?.name || 'default',
             ) as AmqpConnection;
           },
           inject: [RABBIT_CONFIG_TOKEN, AmqpConnectionManager],
@@ -63,7 +60,7 @@ export class RabbitMQModule
         RabbitRpcParamsFactory,
       ],
       exports: [AmqpConnectionManager, AmqpConnection],
-    }
+    },
   )
   implements OnApplicationBootstrap, OnApplicationShutdown
 {
@@ -77,49 +74,30 @@ export class RabbitMQModule
     private readonly externalContextCreator: ExternalContextCreator,
     private readonly rpcParamsFactory: RabbitRpcParamsFactory,
     private readonly connectionManager: AmqpConnectionManager,
-    @Inject(RABBIT_CONFIG_TOKEN) config: RabbitMQConfig
+    @Inject(RABBIT_CONFIG_TOKEN) config: RabbitMQConfig,
   ) {
     super();
     this.logger = config?.logger || new Logger(RabbitMQModule.name);
   }
 
   static async AmqpConnectionFactory(
-    config: RabbitMQConfig
+    config: RabbitMQConfig,
   ): Promise<AmqpConnection | undefined> {
     const logger = config?.logger || new Logger(RabbitMQModule.name);
     if (config == undefined) {
       logger.log(
-        'Rabbitmq config is not provided, skip connection initialization.'
+        'RabbitMQ config not provided, skipping connection initialization.',
       );
       return undefined;
     }
+
+    assertRabbitMqUri(config.uri);
+
     const connection = new AmqpConnection(config);
     this.connectionManager.addConnection(connection);
     await connection.init();
     logger.log('Successfully connected to RabbitMQ');
     return connection;
-  }
-
-  public static build(config: RabbitMQConfig): DynamicModule {
-    const logger = config?.logger || new Logger(RabbitMQModule.name);
-    logger.warn(
-      'build() is deprecated. use forRoot() or forRootAsync() to configure RabbitMQ'
-    );
-    return {
-      module: RabbitMQModule,
-      providers: [
-        {
-          provide: AmqpConnection,
-          useFactory: async (): Promise<AmqpConnection | undefined> => {
-            return config !== undefined
-              ? RabbitMQModule.AmqpConnectionFactory(config)
-              : undefined;
-          },
-        },
-        RabbitRpcParamsFactory,
-      ],
-      exports: [AmqpConnection],
-    };
   }
 
   public static attach(connection: AmqpConnection): DynamicModule {
@@ -148,12 +126,11 @@ export class RabbitMQModule
     connection: AmqpConnection,
     discoveredMethod: DiscoveredMethod,
     config: RabbitHandlerConfig,
-    handler: (...args: any[]) => Promise<any>
+    handler: (...args: any[]) => Promise<any>,
   ) {
     const handlerDisplayName = `${discoveredMethod.parentClass.name}.${
       discoveredMethod.methodName
     } {${config.type}} -> ${
-      // eslint-disable-next-line sonarjs/no-nested-template-literals
       config.queueOptions?.channel ? `${config.queueOptions.channel}::` : ''
     }${config.exchange}::${config.routingKey}::${config.queue || 'amqpgen'}`;
 
@@ -162,24 +139,40 @@ export class RabbitMQModule
       !connection.configuration.enableDirectReplyTo
     ) {
       this.logger.warn(
-        `Direct Reply-To Functionality is disabled. RPC handler ${handlerDisplayName} will not be registered`
+        `Direct Reply-To Functionality is disabled. RPC handler ${handlerDisplayName} will not be registered`,
       );
       return;
     }
 
     this.logger.log(handlerDisplayName);
 
-    return config.type === 'rpc'
-      ? connection.createRpc(handler, config)
-      : connection.createSubscriber(
+    switch (config.type) {
+      case 'rpc':
+        return connection.createRpc(handler, config);
+
+      case 'subscribe':
+        if (config.batchOptions) {
+          return connection.createBatchSubscriber(
+            handler,
+            config,
+            config?.queueOptions?.consumerOptions,
+          );
+        }
+
+        return connection.createSubscriber(
           handler,
           config,
           discoveredMethod.methodName,
-          config?.queueOptions?.consumerOptions
+          config?.queueOptions?.consumerOptions,
         );
+
+      default:
+        throw new Error(
+          `Unable to set up handler ${handlerDisplayName}. Unexpected handler type ${config.type}.`,
+        );
+    }
   }
 
-  // eslint-disable-next-line sonarjs/cognitive-complexity
   public async onApplicationBootstrap() {
     if (RabbitMQModule.bootstrapped) {
       return;
@@ -189,7 +182,7 @@ export class RabbitMQModule
     for (const connection of this.connectionManager.getConnections()) {
       if (!connection.configuration.registerHandlers) {
         this.logger.log(
-          'Skipping RabbitMQ Handlers due to configuration. This application instance will not receive messages over RabbitMQ'
+          'Skipping RabbitMQ Handlers due to configuration. This application instance will not receive messages over RabbitMQ',
         );
 
         continue;
@@ -199,23 +192,23 @@ export class RabbitMQModule
 
       let rabbitMeta =
         await this.discover.providerMethodsWithMetaAtKey<RabbitHandlerConfig>(
-          RABBIT_HANDLER
+          RABBIT_HANDLER,
         );
 
       if (connection.configuration.enableControllerDiscovery) {
         this.logger.log(
-          'Searching for RabbitMQ Handlers in Controllers. You can not use NestJS HTTP-Requests in these controllers!'
+          'Searching for RabbitMQ Handlers in Controllers. You can not use NestJS HTTP-Requests in these controllers!',
         );
         rabbitMeta = rabbitMeta.concat(
           await this.discover.controllerMethodsWithMetaAtKey<RabbitHandlerConfig>(
-            RABBIT_HANDLER
-          )
+            RABBIT_HANDLER,
+          ),
         );
       }
 
       const grouped = groupBy(
         rabbitMeta,
-        (x) => x.discoveredMethod.parentClass.name
+        (x) => x.discoveredMethod.parentClass.name,
       );
 
       const providerKeys = Object.keys(grouped);
@@ -240,11 +233,13 @@ export class RabbitMQModule
               undefined, // contextId
               undefined, // inquirerId
               undefined, // options
-              'rmq' // contextType
+              RABBIT_CONTEXT_TYPE_KEY, // contextType
             );
 
             const moduleHandlerConfigRaw =
-              connection.configuration.handlers[config.name || ''];
+              connection.configuration.handlers[
+                config.name || connection.configuration.defaultHandler || ''
+              ];
 
             const moduleHandlerConfigs = Array.isArray(moduleHandlerConfigRaw)
               ? moduleHandlerConfigRaw
@@ -261,11 +256,11 @@ export class RabbitMQModule
                   connection,
                   discoveredMethod,
                   mergedConfig,
-                  handler
+                  handler,
                 );
-              })
+              }),
             );
-          })
+          }),
         );
       }
     }