+++ packages/stripe/src/stripe.module.ts
@@ -3,7 +3,7 @@ import { createConfigurableDynamicRootModule } from '@golevelup/nestjs-modules';
 import { Logger, Module, OnModuleInit } from '@nestjs/common';
 import { PATH_METADATA } from '@nestjs/common/constants';
 import { ExternalContextCreator } from '@nestjs/core/helpers/external-context-creator';
-import { flatten, groupBy } from 'lodash';
+import { flatten, groupBy, omit } from 'lodash';
 import Stripe from 'stripe';
 import {
   STRIPE_CLIENT_TOKEN,
@@ -35,7 +35,7 @@ export class StripeModule
             Reflect.defineMetadata(
               PATH_METADATA,
               controllerPrefix,
-              StripeWebhookController
+              StripeWebhookController,
             );
             config.webhookConfig?.decorators?.forEach((deco) => {
               deco(StripeWebhookController);
@@ -48,14 +48,13 @@ export class StripeModule
           useFactory: ({
             apiKey,
             typescript = true,
-            apiVersion = '2023-10-16',
-            webhookConfig,
+            apiVersion = '2024-09-30.acacia',
             ...options
           }: StripeModuleConfig): Stripe => {
             return new Stripe(apiKey, {
               typescript,
               apiVersion,
-              ...options,
+              ...omit(options, ['webhookConfig']),
             });
           },
           inject: [STRIPE_MODULE_CONFIG_TOKEN],
@@ -64,7 +63,7 @@ export class StripeModule
         StripePayloadService,
       ],
       exports: [STRIPE_MODULE_CONFIG_TOKEN, STRIPE_CLIENT_TOKEN],
-    }
+    },
   )
   implements OnModuleInit
 {
@@ -74,7 +73,7 @@ export class StripeModule
     private readonly discover: DiscoveryService,
     private readonly externalContextCreator: ExternalContextCreator,
     @InjectStripeModuleConfig()
-    private readonly stripeModuleConfig: StripeModuleConfig
+    private readonly stripeModuleConfig: StripeModuleConfig,
   ) {
     super();
   }
@@ -104,7 +103,7 @@ export class StripeModule
 
     const [stripeWebhookService] = (
       (await this.discover.providersWithMetaAtKey<boolean>(
-        STRIPE_WEBHOOK_SERVICE
+        STRIPE_WEBHOOK_SERVICE,
       )) || []
     ).map((x) => x.discoveredClass.instance);
 
@@ -117,12 +116,12 @@ export class StripeModule
 
     const eventHandlerMeta =
       await this.discover.providerMethodsWithMetaAtKey<string>(
-        STRIPE_WEBHOOK_HANDLER
+        STRIPE_WEBHOOK_HANDLER,
       );
 
     const grouped = groupBy(
       eventHandlerMeta,
-      (x) => x.discoveredMethod.parentClass.name
+      (x) => x.discoveredMethod.parentClass.name,
     );
 
     const webhookHandlers = flatten(
@@ -140,15 +139,17 @@ export class StripeModule
             undefined, // contextId
             undefined, // inquirerId
             undefined, // options
-            'stripe_webhook' // contextType
+            'stripe_webhook', // contextType
           ),
         }));
-      })
+      }),
     );
 
     const handleWebhook = async (webhookEvent: { type: string }) => {
       const { type } = webhookEvent;
-      const handlers = webhookHandlers.filter((x) => x.key === type || x.key === '*');
+      const handlers = webhookHandlers.filter(
+        (x) => x.key === type || x.key === '*',
+      );
 
       if (handlers.length) {
         if (
@@ -156,7 +157,7 @@ export class StripeModule
             ?.logMatchingEventHandlers
         ) {
           this.logger.log(
-            `Received webhook event for ${type}. Forwarding to ${handlers.length} event handlers`
+            `Received webhook event for ${type}. Forwarding to ${handlers.length} event handlers`,
           );
         }
         await Promise.all(handlers.map((x) => x.handler(webhookEvent)));